Move request logic to endpoint class
